Move font traits methods to extension for NSAttributedString and NSMutableAttributedString.
